Add ophan tracking to newsletter preferences
Tracking in Ophan is needed for the newsletters in order to measure
impact of upcoming changes as part of the newsletters OKR.

We have added `data-link-name` attributes to the newsletter checkboxes
and have used the `identityName` as the identifier. This also means that
the `ConsentOptions` model has been updated to include this field.

Co-authored-by: Alex Dufournet <alex.dufournet@theguardian.com>