Update newsletter footer embed src
Following a refactor of how the newsletter embeds will work, there won't
be an iframe hosted at `profile.guardian.co.uk`. Instead I'm moving this
to using `www.theguardian.com` directly, and we'll be implementing some
changes to make sure that things like autofilling email addresses for
logged in users will be implemented in a way that is suitable across
platforms.