Skip JSX transform during TypeScript compilation
We only use the TypeScript compiler for type checking so the actual
JSX to JS transform is handled by Babel. By skipping this we avoid
TypeScript errors from packages that do not use the new JSX transform.
(This is mainly a result of us using an older version of Source and will
be resolved once we update Emotion and Source to the latest versions.)