Override `ecmaVersion` to allow `prefer-const` rule
The Guardian ESLint config is erroneously setting `ecmaVersion` to
`es2020` rather than `2020`. This changes how scoping behaves in some
rules causing `prefer-const` to throw an error in a couple of places
where an assignment to a var does actually happen.

This also changes the behaviour of the `import/namespace` rule so this
has been temporarily added to the overrides pending a fix in a later PR.