Size buttons as `small` where height was not previously set
The custom button styles previously used for this component were
equivalent to a Source button's `small` size. The component's styling
API was being used in some instances to override the button height to
`42px` which would match the default Source button size. In cases where
height was not previously being set we should explicitly set the size
to small.