Add new `recurringSupport` category to grouped product types
This is in support of changes to `members-data-api` to recategorise recurring
contributions and Supporter+ as `recurringSupport` so that they are grouped
together on the Account Overview.

Additionally this adds the required soft opt-ins for the `supporterplus` product
whilst removing these from the grouped product types where they're only required
due to typing. (The `softOptInIDs` property has been retained, but set to a blank array.)

The 'Support The Guardian' section is now only shown for `membership` so the
config for this has been removed from all other products.

Co-authored-by: Andrew Howe-Ely <andrewHEguardian@users.noreply.github.com>
Co-authored-by: James Mockett <jamesmockett@users.noreply.github.com>