Allow product cancellation when visiting cancellation URL directly
Looks up grouped product type using `groupedProductType` property rather
than relying upon `productDetail` as this is not available when visiting
the URL directly as it is passed through as part of the location state.

This also prevents the user being redirected back to the Account
Overview as it was expecting `productDetail` to be part of the router
state, whereas it's already available in the context regardless of how
you arrive at the page.

Co-authored-by: Andrew Howe-Ely <andrewHEguardian@users.noreply.github.com>