Remove unnecessary properties from `GroupedProductType`
Due to `GroupedProductType` extending `ProductType` it includes some
properties that are unnecessary, but must still be included and set to a
default value.

Removing these was complicated by product types and grouped product
types being used interchangeably in some cases. One of the main culprits
was the `mapGroupedToSpecific` function which returns a product type,
but can also return a grouped product type as a fallback. This should
never happen, but if it does it seems more appropriate to throw an error
rather than show potentially incorrect data to users.

Co-authored-by: Richard Bangay <rbangay@users.noreply.github.com>
Co-authored-by: Andrew Howe-Ely <andrewHEguardian@users.noreply.github.com>