feat/refactor: check auth_time in identity middleware
This is a workaround for an Okta behaviour where
some user types won't be asked to reauthenticate
even when max_age is set on the OAuth request. We
check the auth_time of the ID token ourselves
in the identity middleware, ensuring that it is
within max_age. If it isn't, we revoke the access
token, clear the token cookies, and redirect the
browser to the Gateway /reauthenticate endpoint,
which always shows a sign-in form.